% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dd_filters.R
\name{apply_dd_filters}
\alias{apply_dd_filters}
\title{Apply filters to prepare data for degree-day calculation}
\usage{
apply_dd_filters(
  dat,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 540,
  full_season = TRUE,
  heat_threshold = 18,
  n_hours = 24
)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{trend_threshold}{The threshold for "trending up". Default is
\code{trend_threshold = 4}. The last observation above
\code{trend_threshold} that does not return below \code{trend_threshold}
triggers the beginning of the growing season for each \code{DEPTH} and
group in \code{...}.}

\item{superchill_threshold}{The threshold for "superchill". Default is
\code{superchill_threshold = -0.7}. The first observation below
\code{superchill_threshold} triggers the end of the growing season for each
group in \code{DEPTH} and \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}. Default
is \code{max_season = 540} days (~18 months). Note: units are days because
adding 18 months to August 30 or August 31 results in \code{NA} (because
February 30 and February 31 are not real dates).}

\item{full_season}{Logical argument. The default, \code{full_season = TRUE}
will only return data for groups with a full season of data (e.g.,
\code{START_SEASON} and \code{END_SEASON} as returned by
\code{identify_growing_seasons(dat, full_season = TRUE)} are both NOT
\code{NA}).

If \code{full_season = FALSE}, data for all groups will be returned;
the \code{START_SEASON} and \code{END_SEASON} dates default to the first
and last timestamp in \code{...} and \code{DEPTH}.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns \code{dat} filtered for days that will be used to calculate
degree-days at \code{DEPTH}. Includes observations that start after the
4-degree trending up threshold and end 1 minute before the first
observation of superchill for each season; observations that occur during a
heat stress events are removed.

An additional column \code{SEASON} is included to label the growing seasons
("S1", "S2", ...). Some observations may be duplicated in consecutive
seasons.
}
\description{
Apply filters to prepare data for degree-day calculation
}
