% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_mode}
\alias{calc_mode}
\title{Extract the most common value in a vector}
\usage{
calc_mode(v)
}
\arguments{
\item{v}{Vector for which to extract the mode.}
}
\description{
Extract the most common value in a vector
}
