% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_degree_days.R
\name{calculate_degree_days}
\alias{calculate_degree_days}
\title{Calculates temperature in degree-days}
\usage{
calculate_degree_days(
  dat,
  ...,
  start.date = min(dat$TIMESTAMP),
  end.date = max(dat$TIMESTAMP)
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Columns in \code{dat} to use for grouping in
\code{dplyr::group_by()}, e.g., \code{DEPTH}.}

\item{start.date}{First day to include in the calculation. If
\code{start.date} does not include a time, then the start time is assumed
to be midnight. Accepted orders for \code{start.date} are: "ymd IMS p",
"Ymd IMS p", "Ymd HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd".
Default is the first TIMESTAMP in \code{dat}.}

\item{end.date}{Last day to include in the calculation. If \code{end.date}
does not include a time, then the end time is assumed to be "23:59:59".
Accepted orders for \code{end.date} are: "ymd IMS p", "Ymd IMS p", "Ymd
HM", "Ymd HMS", "dmY HM", "dmY HMS", "Ymd", "ymd". Default is the last
TIMESTAMP in \code{dat}.}
}
\value{
Returns a tibble with at least five columns: \code{PERIOD} (start and
end date used to calculate mean temperature and number of days),
\code{n_DAYS} (the number of days, \emph{n}, used in the calculation),
\code{n_OBSERVATIONS} (the number of observations used to calculate the
average temperature), \code{AVG_TEMPERATURE} (the average temperature in
the time period), \code{DEGREE_DAYS} (degree-days, the product of
\code{AVG_TEMPERATURE} and \code{n_DAYS}). Additional columns are returned
for each grouping variable in \code{...}.
}
\description{
Calculates temperature in degree-days
}
\details{
Degree-days = average temperature over \emph{n} days * \emph{n} days

\emph{n} is calculated as the number of unique dates in the TIMESTAMP
column. This means that days without data (e.g. due to sensor mal-function
or deliberately filtered out) will NOT be included in the degree-day
calculation.For example: a sensor is deployed from September 1, 2021 to
September 14, 2021 (14 days).

The temperature for several observations on September 3 was > 20 degree
Celcius, and so that day is filtered out of the analysis. In the degree-day
calculation \emph{n} = 13 days.
}
\examples{
data(string_data)

# degree-days averaged over all depths and the whole time series
calculate_degree_days(string_data)

# degree-days by DEPTH for whole time series
calculate_degree_days(string_data, DEPTH)

# degree-days by DEPTH and month from July 1 to September 30
calculate_degree_days(string_data, DEPTH, months(TIMESTAMP),
start.date = "2019-07-01", end.date = "2019-09-30")
}
\author{
Danielle Dempsey
}
\concept{calculate}
