% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_functions.R
\name{calculate_degree_days_TGC}
\alias{calculate_degree_days_TGC}
\title{Use TGC model to calculate degree days required to grow from initial to final
weight}
\usage{
calculate_degree_days_TGC(initial_weight, final_weight, tgc)
}
\arguments{
\item{initial_weight}{Initial weight of the fish}

\item{final_weight}{Final weight of the fish}

\item{tgc}{Thermal growth coefficient}
}
\value{
Returns a single value
}
\description{
Use TGC model to calculate degree days required to grow from initial to final
weight
}
