% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_event_duration.R
\name{calculate_event_duration}
\alias{calculate_event_duration}
\title{Calculate the duration of time VALUE exceeds a threshold (above or below)}
\usage{
calculate_event_duration(
  dat,
  threshold,
  ...,
  less_than = TRUE,
  summary_table = TRUE
)
}
\arguments{
\item{dat}{Dataframe that must include columns VALUE (numeric) and TIMESTAMP
(POSIXct), and a column for each grouping variable in \code{...}.}

\item{threshold}{Value to cross to trigger exceedance (exclusive)}

\item{...}{Names of columns that should be used as grouping variables (not
quoted), e.g. \code{DEPTH, YEAR}.}

\item{less_than}{Logical argument. If \code{TRUE} (the default), function
will identify events when VALUE was BELOW \code{threshold}; if
\code{FALSE}, will identify events when VALUE was ABOVE the
\code{threshold}.}

\item{summary_table}{Logical argument. If \code{TRUE} (the default), will
return summary table with maximum event duration for each grouping
variable. If \code{FALSE}, will return table with all exceedance events for
each grouping variable.}
}
\value{
Returns summary table with maximum event duration for each grouping
variable OR a table with all exceedance events for each grouping variable.
}
\description{
Calculate the duration of time VALUE exceeds a threshold (above or below)
}
\details{
No row will be returned for level of grouping variable for which
there were no exceedance events. e.g., if there was no superchill at 5 m,
there will be no row for 5 m in dat.out.

Do not need to arrange by variables in ... (grouping takes care of that).
}
