% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_functions.R
\name{calculate_final_weight_TGC}
\alias{calculate_final_weight_TGC}
\title{Calculate final body weight using TGC model}
\usage{
calculate_final_weight_TGC(initial_weight, tgc, degree_days)
}
\arguments{
\item{initial_weight}{Initial weight of the fish}

\item{tgc}{Thermal growth coefficient}

\item{degree_days}{Number of degree days experienced}
}
\value{
Returns a single value
}
\description{
Calculate final body weight using TGC model
}
\details{
final_weight = (initial_weight^(1/3) + (tgc/1000)*degree_days)^3
}
