% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_functions.R
\name{calculate_initial_weight_TGC}
\alias{calculate_initial_weight_TGC}
\title{Calculate initial body weight using TGC model}
\usage{
calculate_initial_weight_TGC(final_weight, tgc, degree_days)
}
\arguments{
\item{final_weight}{Final weight of the fish}

\item{tgc}{Thermal growth coefficient}

\item{degree_days}{Number of degree days experienced}
}
\value{
Returns a single value
}
\description{
Calculate initial body weight using TGC model
}
