% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_data_gaps.R
\name{check_for_data_gaps}
\alias{check_for_data_gaps}
\title{Check for substantial gaps in the time series}
\usage{
check_for_data_gaps(dat, ..., gap_length = 2, gap_warning = 6)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{DEPTH}.}

\item{gap_length}{The length of time in hours to consider a sampling gap.
Default is \code{gap_length = 2} hours, which is twice the sampling
interval of the least frequent sensor.}

\item{gap_warning}{The length of time in hours to consider a substantial
sampling gap. A warning will be printed if any gaps exceed this duration.
Default is \code{gap_warning = 6} hours.}
}
\value{
Returns summary table with \code{GAP_START},\code{GAP_LENGTH_DAYS},
and \code{GAP_LENGTH_HOURS} for each grouping variable for all intervals
between observations that exceed \code{gap_length}.

A warning will be displayed if any intervals exceed \code{gap_warning}.
}
\description{
Check for substantial gaps in the time series
}
