% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_growing_days.R
\name{count_growing_days}
\alias{count_growing_days}
\title{Count number of days that were suitable for growth}
\usage{
count_growing_days(
  dat,
  ...,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 18,
  heat_threshold = 18,
  n_hours = 24
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{SEASON} and \code{DEPTH}.}

\item{trend_threshold}{The threshold for "trending up". Default is
\code{trend_threshold = 4}. The last observation above
\code{trend_threshold} that does not return below \code{trend_threshold}
triggers the beginning of the growing season for each \code{DEPTH} and
group in \code{...}.}

\item{superchill_threshold}{The threshold for "superchill". Default is
\code{superchill_threshold = -0.7}. The first observation below
\code{superchill_threshold} triggers the end of the growing season for each
group in \code{DEPTH} and \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns a tibble with columns: \code{...}, \code{DEPTH},
\code{SEASON}, \code{START_SEASON} (minimum TIMESTAMP for each group),
\code{END_SEASON} (maximum TIMESTAMP for each group), \code{TOTAL_DAYS}
(\code{difftime(END_SEASON, START_SEASON, units = "days")}),
\code{n_filtered_days} (calculated from
\code{identify_heat_stress_events}), and \code{n_growing_days} (TOTAL_DAYS
\itemize{
\item n_filtered_days).
}
}
\description{
Count number of days that were suitable for growth
}
\details{
Days suitable for growth are days that remain after applying season
and heat stress filters (\code{apply_dd_filters}).

Results are automatically grouped by \code{SEASON} and \code{DEPTH}.
}
\examples{
data(string_data)
string_data <- string_data[which(string_data$VARIABLE == "Temperature"), ]
count_growing_days(string_data)
}
