% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_station_label.R
\name{create_station_label}
\alias{create_station_label}
\title{Assign a label in the form Station Name Long Season}
\usage{
create_station_label(dat, short, medium, long, sep = "\\n")
}
\arguments{
\item{dat}{Data frame with at least one column named \code{STATION}.}

\item{short}{Station name for the short season example (character string).}

\item{medium}{Station name for the medium season example (character string).}

\item{long}{Station name for the long season example (character string).}

\item{sep}{Label separator. Default is a line break.}
}
\value{
Returns \code{dat} with additional column \code{LABEL}.
}
\description{
Assign a label in the form Station Name Long Season
}
