% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_heat_stress_events.R
\name{filter_heat_stress_events}
\alias{filter_heat_stress_events}
\title{Filters out observations that occur during heat stress events}
\usage{
filter_heat_stress_events(dat, heat_threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{heat_threshold}{The threshold to trigger heat stress interval
(inclusive. Default is \code{heat_threshold = 18}).}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns dat, filtered to remove observations that occur during
heat stress events.
}
\description{
Filters out observations that occur during heat stress events
}
\details{
Every observation that exceeds the threshold has a corresponding
heat stress interval (\code{TIMESTAMP} of the observation + n_hours).
Intervals may overlap with one or more other intervals.

Heat stress events are denoted by the beginning and end of overlapping
intervals for each depth. Heat stress events do not overlap, but they may
end and start on consecutive timestamps.
}
\seealso{
Other heat stress: 
\code{\link{identify_heat_stress_intervals}()}
}
\concept{heat stress}
