% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_heat_stress_events_single.R
\name{filter_out_heat_stress_events_single}
\alias{filter_out_heat_stress_events_single}
\title{Filter out observations that occur during heat stress events}
\usage{
filter_out_heat_stress_events_single(dat, heat_threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns \code{dat}, filtered to remove observations that occur during
heat stress events.
}
\description{
Filter out observations that occur during heat stress events
}
\details{
Filters out observations that occur during heat stress events, as
defined by \code{identify_heat_stress_events()}.

This function only works for a single \code{STATION}. For multiple
\code{STATION}s, use \code{filter_out_heat_stress_events()}.

Every observation that exceeds the threshold has a corresponding heat
stress interval (\code{TIMESTAMP} of the observation + n_hours). Intervals
may overlap with one or more other intervals.

Heat stress events are denoted by the beginning and end of overlapping
intervals for each depth. Heat stress events do not overlap, but they may
end and start on consecutive timestamps.
}
