% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_days_above_threshold.R
\name{identify_days_above_threshold}
\alias{identify_days_above_threshold}
\title{Identifies days during which threshold is exceeded}
\usage{
identify_days_above_threshold(
  dat,
  threshold = 20,
  min_exceedance = 1,
  n_extra_days = 0
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. Other
columns will be ignored.}

\item{threshold}{The threshold to trigger filtering data (inclusive. Default
is \code{threshold = 20}).}

\item{min_exceedance}{The minimum number of observations in a day to trigger
filtering data (inclusive). Default is \code{min.exceedence = 1}.}

\item{n_extra_days}{Number of additional days to filter out when
\code{threshold} and \code{min_exceedence} are exceeded. If
\code{n_extra_days = 0} (the default), then all values from the day the
threshold is exceeded will be filtered out. If \code{n_extra_days = 1},
then all values from the day the threshold is exceeded and the following
day will be filtered out, etc.}
}
\value{
Returns table of dates that should be filtered based on threshold,
min_exceedance, and n_extra_days. Three columns: DEPTH, DATE, and n_obs
(the number of observations that exceeded the threshold at the DEPTH and
DATE).
}
\description{
Identifies days when the value exceeds a given threshold for a
given number of observations. Option to identify additional days after the
threshold is exceeded.

\code{dat} must only include data for one variable.
}
