% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_first_superchill.R
\name{identify_first_superchill}
\alias{identify_first_superchill}
\title{Identifies when VALUE first crosses below a lower threshold}
\usage{
identify_first_superchill(dat, ..., superchill_threshold = -0.7)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{DEPTH}.}

\item{superchill_threshold}{Default is \code{superchill_threshold = -0.7}.
The first observation below \code{superchill_threshold} triggers the end of
the growing season for each group in \code{...}.}
}
\value{
Returns the TIMESTAMP (for each group in \code{...}) for the first
time VALUE goes below \code{superchill_threshold}.

This TIMESTAMP is passed to \code{identify_growing_seasons()} to denote the
end of the growing season.

No row will be returned for groups for which VALUE did not cross
\code{superchill_threshold}.
}
\description{
Identifies when VALUE first crosses below a lower threshold
}
\details{
Identifies TIMESTAMP when VALUE first crosses below threshold for
each group in \code{...}.

If the VALUE does not cross the threshold for any groups, a dataframe with
0 rows will be returned.
}
