% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_first_superchill_days.R
\name{identify_first_superchill_days}
\alias{identify_first_superchill_days}
\title{Identifies when VALUE first crosses below a lower threshold}
\usage{
identify_first_superchill_days(dat, ..., superchill_threshold = -0.7)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Columns in \code{dat} to use for grouping in
\code{dplyr::group_by()}, e.g., \code{YEAR, DEPTH}.}

\item{superchill_threshold}{Default is \code{lower_threshold = -0.7}.}
}
\value{
Returns the TIMESTAMP (for each group in \code{...}) for the first
time VALUE goes below \code{superchill_threshold}.

No row will be returned for groups for which VALUE did not cross
\code{superchill_threshold}.
}
\description{
Identifies when VALUE first crosses below a lower threshold
}
\details{
Identifies timestamp when value first crosses below threshold for
each group in \code{...}.
}
