% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_growing_seasons.R
\name{identify_growing_seasons}
\alias{identify_growing_seasons}
\title{Identify start and end of growing seasons for each year}
\usage{
identify_growing_seasons(
  dat,
  ...,
  full_season = TRUE,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 540
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{DEPTH}.}

\item{full_season}{Logical argument. The default, \code{full_season = TRUE}
will return \code{NA} for \code{START_SEASON} and \code{END_SEASON} that
occur outside of the date range of the data series. For example, if the
data series starts when temperature is already above
\code{trend_threshold}, \code{START_SEASON = NA}; if the season length is
determined by \code{max_season}, which results in a date after the last
observation, \code{END_SEASON = NA}. If \code{full_season = FALSE}, the
\code{START_SEASON} and \code{END_SEASON} dates default to the first and
last timestamp in \code{...} and \code{DEPTH}.}

\item{trend_threshold}{The threshold for "trending up". Default is
\code{trend_threshold = 4}. The last observation above
\code{trend_threshold} that does not return below \code{trend_threshold}
triggers the beginning of the growing season for each \code{DEPTH} and
group in \code{...}.}

\item{superchill_threshold}{The threshold for "superchill". Default is
\code{superchill_threshold = -0.7}. The first observation below
\code{superchill_threshold} triggers the end of the growing season for each
group in \code{DEPTH} and \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}. Default
is \code{max_season = 540} days (~18 months). Note: units are days because
adding 18 months to August 30 or August 31 results in \code{NA} (because
February 30 and February 31 are not real dates).}
}
\value{
Returns a tibble with the \code{START_SEASON}, \code{END_SEASON},
\code{SEASON_DAYS} (length of season in days), and  \code{SEASON_MONTHS}
(season length in months; assumes 1 month = 30 days) for each group in
\code{DEPTH} and group in \code{...}.
}
\description{
Identify start and end of growing seasons for each year
}
\details{
The growing season starts when the temperature crosses
\code{trend_threshold} and does not return below \code{trend_threshold}
(e.g., 4-degrees trending up) and ends 1 minute before the first
observation of \code{superchill_threshold}.

The function looks for \code{START_SEASON} in January to August to avoid
temperature drops caused by hurricanes (typically in September and October)
and Winter cooling (November, December).

The function looks for \code{END_SEASON} in January to May of the next year
(spring superchill).

If the time series begins when the temperature is above
\code{trend_threshold}, \code{START_SEASON} is assigned the minimum
\code{TIMESTAMP} for the group in \code{filter_growing_season()}.

If the time series does not go below \code{superchill_threshold}, the
growing season is assumed to be \code{max_season} months long.
\code{END_SEASON} is assigned the timestamp \code{START_SEASON} +
\code{days(max_season)}.

Growing seasons may overlap if \code{max_season} is used to determine
\code{END_SEASON}.
}
