% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_heat_stress_events.R
\name{identify_heat_stress_events}
\alias{identify_heat_stress_events}
\title{Identifies start and end of non-overlapping heat stress events}
\usage{
identify_heat_stress_events(dat, threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{threshold}{The threshold to trigger heat stress interval (inclusive.
Default is \code{threshold = 18}).}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns a dataframe with four columns: \code{DEPTH}, \code{event_id},
\code{stress_start} and \code{stress_end}.

Note: Events for a given depth do not overlap, but \code{event_end} and
\code{event_start} may be consecutive timestamps.
}
\description{
Identifies start and end of non-overlapping heat stress events
}
\details{
Every observation that exceeds the threshold has a corresponding
heat stress interval (\code{TIMESTAMP} of the observation + n_hours).
Intervals may overlap with one or more other intervals.

Heat stress events are denoted by the beginning and end of overlapping
intervals for each depth. Heat stress events do not overlap, but they may
end and start on consecutive timestamps.
}
