% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_heat_stress_intervals.R
\name{identify_heat_stress_intervals}
\alias{identify_heat_stress_intervals}
\title{Identify n-hour intervals beginning when a threshold is exceeded}
\usage{
identify_heat_stress_intervals(dat, ..., heat_threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{DEPTH}.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns a data frame with columns: \code{...},  \code{DEPTH},
\code{interval_start} and \code{interval_end}. \code{interval_start}
indicates the beginning of each heat stress interval, i.e, the
\code{TIMESTAMP} of all observations where \code{VALUE} is greater than or
equal to \code{heat_threshold}. \code{interval_end} identifies the end of
the heat stress interval, i.e., \code{interval_start} + n_hours.

Note: intervals may overlap with previous interval(s).
}
\description{
Identify n-hour intervals beginning when a threshold is exceeded
}
\examples{
data(string_data)

dat <- string_data[which(string_data$VARIABLE == "Temperature"), ]
stress_intervals <- identify_heat_stress_intervals(dat)
}
\seealso{
Other heat stress: 
\code{\link{filter_out_heat_stress_events}()}
}
\concept{heat stress}
