% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_trending_up.R
\name{identify_trending_up}
\alias{identify_trending_up}
\title{Identifies when VALUE begins trending up}
\usage{
identify_trending_up(dat, ..., trend_threshold = 4)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Columns in \code{dat} to use for grouping in
\code{dplyr::group_by()}, e.g., \code{DEPTH}.}

\item{trend_threshold}{Default is \code{trend_threshold = 4}. The last
observation above \code{trend_threshold} that does not return below
\code{trend_threshold} triggers the beginning of the growing season for
each group in \code{...}.}
}
\value{
Returns the TIMESTAMP (for each group in \code{...}) for the final
time VALUE exceeds \code{trend_threshold} and does not return below
\code{trend_threshold}.

This TIMESTAMP is passed to \code{identify_growing_seasons()} to denote the
start of the growing season.

No row will be returned for groups for which VALUE did not cross
\code{trend_threshold}.
}
\description{
Identifies when VALUE begins trending up
}
\details{
Identifies timestamp when value crosses threshold and does not
return below the threshold, for each group in \code{...}.
}
\examples{
data("string_data")

string_data <- string_data[which(string_data$VARIABLE == "Temperature"), ]
trend_up <- identify_trending_up(string_data, trend_threshold = 4, DEPTH)
}
