% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_degree_days.R
\name{plot_degree_days}
\alias{plot_degree_days}
\title{Plot degree days at depth for each season start date}
\usage{
plot_degree_days(dd_table, facet_var = NULL, ncol = 1, nrow = NULL)
}
\arguments{
\item{dd_table}{Dataframe with degree-day data, as exported from
\code{count_degree_days}. Must include columns \code{DEPTH},
\code{START_SEASON}, \code{n_degree_days}.}

\item{facet_var}{Variable(s) defining faceting groups. Variables must be
column(s) in \code{dat}). For a single facet variable: \code{facet_var =
  "SEASON"}. For more than one facet variables: \code{facet_var = "SEASON +
  DEPTH"}. Default is \code{facet_var = NULL}.}

\item{ncol}{Number of columns for faceted figure. Default is \code{ncol = 1}.}

\item{nrow}{Number of rows for faceted figure. Default is \code{nrow = NULL}.}
}
\value{
ggplot object
}
\description{
Plot degree days at depth for each season start date
}
