% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filtered_data.R
\name{plot_filtered_data}
\alias{plot_filtered_data}
\title{Plot filtered data with removed data greyed out}
\usage{
plot_filtered_data(
  dat,
  dat_filtered,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  heat_threshold = 18
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{dat_filtered}{Filtered data, i.e. the result of
\code{apply_dd_filters(dat)}.}

\item{trend_threshold}{Default is \code{trend_threshold = 4}. The last
observation above \code{trend_threshold} that does not return below
\code{trend_threshold} triggers the beginning of the growing season for
each group in \code{...}.}

\item{superchill_threshold}{Default is \code{superchill_threshold = -0.7}.
The first observation below \code{superchill_threshold} triggers the end of
the growing season for each group in \code{...}.}

\item{heat_threshold}{The threshold to trigger heat stress interval
(inclusive. Default is \code{heat_threshold = 18}).}
}
\value{
Returns a ggplot object.
}
\description{
Plot filtered data with removed data greyed out
}
