% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_data.R
\docType{data}
\name{string_data}
\alias{string_data}
\title{Example temperature data}
\format{
This example data is a dataframe with 89370 rows and 4 variables:
\describe{ \item{STATION}{Station where measurement was recorded}
\item{TIMESTAMP}{Timestamp of the observation} \item{DEPTH}{Depth of the
sensor} \item{VALUE}{Value of temperature measurement} }
}
\source{
Data from CMAR's Coastal Monitoring Program
}
\usage{
string_data
}
\description{
Example temperature data
}
\keyword{datasets}
