% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_functions.R
\name{TGC_calculate_coefficent}
\alias{TGC_calculate_coefficent}
\title{Calculate the thermal growth coefficient}
\usage{
TGC_calculate_coefficent(initial_weight, final_weight, average_temp, n_days)
}
\arguments{
\item{initial_weight}{A vector of initial weight(s) of the fish in grams.}

\item{final_weight}{A vector of final weight(s) of the fish in grams.}

\item{average_temp}{Average temperature of the growing days.}

\item{n_days}{Number of growing days.}
}
\value{
Returns a table with columns \code{INITIAL_WEIGHT},
\code{FINAL_WEIGHT}, \code{degree_days}, and \code{TGC_COEF}.
}
\description{
Calculate the thermal growth coefficient
}
