% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dd_filters.R
\name{apply_dd_filters}
\alias{apply_dd_filters}
\title{Apply filters to prepare data for degree-day calculation}
\usage{
apply_dd_filters(
  dat,
  ...,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 18,
  heat_threshold = 18,
  n_hours = 24
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{...}{Columns in \code{dat} to use for grouping in
\code{dplyr::group_by()}, e.g., \code{DEPTH}.}

\item{trend_threshold}{Default is \code{trend_threshold = 4}. The last
observation above \code{trend_threshold} that does not return below
\code{trend_threshold} triggers the beginning of the growing season for
each group in \code{...}.}

\item{superchill_threshold}{Default is \code{superchill_threshold = -0.7}.
The first observation below \code{superchill_threshold} triggers the end of
the growing season for each group in \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}.}

\item{heat_threshold}{The threshold to trigger heat stress interval
(inclusive. Default is \code{heat_threshold = 18}).}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns \code{dat} filtered for days that will be used to calculate
degree-days at DEPTH. Includes observations that start after the 4-degree
trending up threshold and end 1 minute before the first observation
superchill for each season; observations that occur during a heat stress
events are removed.

An additional column \code{SEASON} is included to label the growing seasons
("S1", "S2", ...). Some observations may be duplicated in consecutive
seasons.
}
\description{
Apply filters to prepare data for degree-day calculation
}
