% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_event_duration.R
\name{calculate_event_duration}
\alias{calculate_event_duration}
\title{Calculate the duration of time VALUE exceeds a threshold (above or below)}
\usage{
calculate_event_duration(
  dat,
  threshold,
  ...,
  less_than = TRUE,
  summary_table = TRUE
)
}
\arguments{
\item{dat}{Data frame with columns \code{VALUE} (numeric), \code{TIMESTAMP}
(POSIXct), and a column for each grouping variable in \code{...}.}

\item{threshold}{Threshold value that triggers start of event. Event starts
when \code{VALUE} is greater than (less than) OR equal to \code{threshold}.}

\item{...}{Grouping variables. Must be name(s) of column(s) in \code{dat}.}

\item{less_than}{Logical argument. If \code{TRUE}, function identifies events
when \code{VALUE} was less than or equal to \code{threshold}; if
\code{FALSE}, the function identifies events when \code{VALUE} was greater
than or equal to \code{threshold}.}

\item{summary_table}{Logical argument. If \code{FALSE}, function will return
table with all events for each grouping variable. If \code{TRUE}, function
will return summary table with maximum event duration for each grouping
variable.}
}
\value{
Returns a data frame.
}
\description{
Calculate the duration of time VALUE exceeds a threshold (above or below)
}
