% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_degree_days.R
\name{count_degree_days}
\alias{count_degree_days}
\title{Count number of degree-days}
\usage{
count_degree_days(dat_filt, ..., growing_days)
}
\arguments{
\item{dat_filt}{Dataframe that has been filtered by
\code{apply_dd_filters()}. Includes columns: \code{...}, \code{SEASON},
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. May also include columns with grouping
variables passed to \code{...}. Other columns will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{SEASON} and \code{DEPTH}.}

\item{growing_days}{Data.frame for each group}
}
\value{
Returns a tibble with columns:
}
\description{
Count number of degree-days
}
\details{
Degree-days = average temperature over \emph{n} days * \emph{n} days

\emph{n} is the number of days suitable for growth (from
\code{count_growing_days()}).

Results are grouped by \code{SEASON}, \code{DEPTH}, and \code{...}.

*\strong{Could put count_growing_days and apply_dd_filters in here}
}
\author{
Danielle Dempsey
}
\concept{calculate}
