% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_growing_days.R
\name{count_growing_days}
\alias{count_growing_days}
\title{Count number of days that were suitable for growth}
\usage{
count_growing_days(dat, ..., heat_threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{...}{Additional grouping variables (beside \code{DEPTH}).}

\item{heat_threshold}{The threshold to trigger heat stress interval
(inclusive. Default is \code{heat_threshold = 18}).}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns a tibble with columns: DEPTH, START_SEASON (minimum TIMESTAMP
for each group), END_SEASON (maximum TIMESTAMP for each group), TOTAL_DAYS
\code{difftime(END_SEASON, START_SEASON, units = "days")}, n_filtered_days
(calculated from \code{identify_heat_stress_events}), and n_growing_days
(TOTAL_DAYS - n_filtered_days).
}
\description{
Count number of days that were suitable for growth
}
