% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_growing_days.R
\name{count_growing_days}
\alias{count_growing_days}
\title{Count number of days that were suitable for growth}
\usage{
count_growing_days(
  dat,
  ...,
  heat_threshold = 18,
  n_hours = 24,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 540,
  full_season = TRUE,
  rm_gap_days = FALSE,
  gap_length = 2,
  gap_warning = 6,
  quiet = TRUE
)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{...}{Additional columns in \code{dat} to use as grouping variables.
Results are automatically grouped by \code{SEASON} and \code{DEPTH}.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}

\item{trend_threshold}{The threshold for "trending up". Default is
\code{trend_threshold = 4}. The last observation above
\code{trend_threshold} that does not return below \code{trend_threshold}
triggers the beginning of the growing season for each \code{DEPTH} and
group in \code{...}.}

\item{superchill_threshold}{The threshold for "superchill". Default is
\code{superchill_threshold = -0.7}. The first observation below
\code{superchill_threshold} triggers the end of the growing season for each
group in \code{DEPTH} and \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}. Default
is \code{max_season = 540} days (~18 months). Note: units are days because
adding 18 months to August 30 or August 31 results in \code{NA} (because
February 30 and February 31 are not real dates).}

\item{full_season}{Logical argument. The default, \code{full_season = TRUE}
will only return data for groups with a full season of data (e.g.,
\code{START_SEASON} and \code{END_SEASON} as returned by
\code{identify_growing_seasons(dat, full_season = TRUE)} are both NOT
\code{NA}).

If \code{full_season = FALSE}, data for all groups will be returned;
the \code{START_SEASON} and \code{END_SEASON} dates default to the first
and last timestamp in \code{...} and \code{DEPTH}.}

\item{rm_gap_days}{Logical argument indicating whether to remove days with
data gaps when counting n_growing_days.}

\item{gap_length}{The length of time in hours to consider a sampling gap.
Default is gap_length = 2 hours, which is twice the sampling interval of
the least frequent sensor. The total sample gap length (sum of all sample
gaps) is subtracted from the number of stocked days to calculate the number
of growing days.}

\item{gap_warning}{The length of time in hours to consider a substantial
sampling gap. A warning can be printed if any gaps exceed this duration.
Default is \code{gap_warning = 6} hours.}

\item{quiet}{Logical argument to suppress Warning message when gap lengths
exceed \code{gap_warning}. If \code{quiet = TRUE}, the Warning will not be
printed.}
}
\value{
Returns a tibble with columns: \code{...}, \code{DEPTH},
\code{SEASON}, \code{START_SEASON} (minimum TIMESTAMP for each group),
\code{END_SEASON} (maximum TIMESTAMP for each group), \code{STOCKED_DAYS}
(\code{difftime(END_SEASON, START_SEASON, units = "days")}),
\code{n_filtered_days} (calculated from
\code{identify_heat_stress_events}), and \code{n_growing_days}
(STOCKED_DAYS - n_filtered_days).
}
\description{
Count number of days that were suitable for growth
}
\details{
Days suitable for growth are days that remain after applying season
and heat stress filters. Option to remove days with data gaps (otherwise
the assumption is that the temperature on these days is the average
temperature of the data series).

Results are automatically grouped by \code{SEASON} and \code{DEPTH}.

Runs filter_in_growing_seasons() if no SEASON column in the data. If there
is not a full season of data, make sure to set argument full_season = FALSE
or add a SEASON column to dat.
}
