% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_growing_seasons.R
\name{filter_growing_seasons}
\alias{filter_growing_seasons}
\title{Keeps observations that occur during growing seasons}
\usage{
filter_growing_seasons(
  dat,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  max_season = 18
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (must be
possible to convert to a Date object), \code{DEPTH}, and \code{VALUE}. If
column \code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{trend_threshold}{Default is \code{trend_threshold = 4}. The last
observation above \code{trend_threshold} that does not return below
\code{trend_threshold} triggers the beginning of the growing season for
each group in \code{...}.}

\item{superchill_threshold}{Default is \code{superchill_threshold = -0.7}.
The first observation below \code{superchill_threshold} triggers the end of
the growing season for each group in \code{...}.}

\item{max_season}{The length of the growing season in months for groups that
have no temperature observations below \code{superchill_threshold}.}
}
\value{
Returns dat, filtered for observations that occur during growing
seasons identified in \code{identify_growing_seasons}, with an additional
column \code{SEASON} to label the seasons ("S1", "S2", ...). Some
observations may be duplicated in consecutive seasons.
}
\description{
Keeps observations that occur during growing seasons
}
\details{
for ONE STATION
}
