% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_heat_stress_intervals.R
\name{identify_heat_stress_intervals}
\alias{identify_heat_stress_intervals}
\title{Identifies n-hour intervals beginning when threshold is exceeded}
\usage{
identify_heat_stress_intervals(dat, threshold = 18, n_hours = 24)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{threshold}{The threshold to trigger heat stress interval (inclusive.
Default is \code{threshold = 18}).}

\item{n_hours}{Length of heat stress interval in hours (default is
\code{n_hours = 24}).}
}
\value{
Returns a dataframe with three columns: \code{DEPTH},
\code{interval_start} and \code{interval_end}. \code{interval_start}
indicates the beginning of each heat stress interval, i.e, the
\code{TIMESTAMP} of all observations where \code{VALUE} is greater than or
equal to \code{threshold}. \code{interval_end} identifies the end of the
heat stress interval, i.e., \code{interval_start} + n_hours.

Note: intervals may overlap with previous interval(s).
}
\description{
Identifies n-hour intervals beginning when threshold is exceeded
}
\examples{

data(string_data)

dat <- string_data[which(string_data$VARIABLE == "Temperature"), ]
stress_intervals <- identify_heat_stress_intervals(dat)
}
\seealso{
Other heat stress: 
\code{\link{filter_heat_stress_events}()}
}
\concept{heat stress}
