% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filtered_data.R
\name{plot_filtered_data}
\alias{plot_filtered_data}
\title{Plot filtered data with removed data greyed out}
\usage{
plot_filtered_data(
  dat,
  dat_filtered,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  heat_threshold = 18,
  colour_palette = NULL,
  legend_drop = FALSE,
  legend_position = "right",
  ylims = NULL,
  date_axis_name = "Date",
  date_breaks_major = NULL,
  date_breaks_minor = NULL,
  date_labels_format = NULL,
  plotly_friendly = FALSE,
  alpha = 1
)
}
\arguments{
\item{dat}{Data frame with at least three columns: \code{TIMESTAMP} (must be
possible to convert to POSIXct), \code{DEPTH}, and \code{VALUE}. If column
\code{VARIABLE} is included, it must have one unique entry. May also
include columns with grouping variables passed to \code{...}. Other columns
will be ignored.}

\item{dat_filtered}{Filtered data, i.e. the result of
\code{apply_dd_filters(dat)}.}

\item{trend_threshold}{The threshold for "trending up". Default is
\code{trend_threshold = 4}. The last observation above
\code{trend_threshold} that does not return below \code{trend_threshold}
triggers the beginning of the growing season for each \code{DEPTH} and
group in \code{...}.}

\item{superchill_threshold}{The threshold for "superchill". Default is
\code{superchill_threshold = -0.7}. The first observation below
\code{superchill_threshold} triggers the end of the growing season for each
group in \code{DEPTH} and \code{...}.}

\item{heat_threshold}{The threshold for heat stress. Default is
\code{heat_threshold = 18}). Every observation above \code{heat_threshold}
triggers an \code{n-hour} heat stress interval.}

\item{colour_palette}{Optional vector of hex colors onto which \code{DEPTH}
will be mapped. If \code{colour_palette = NULL} (the default), the reverse
viridis colour palette will be used (option D).}

\item{legend_drop}{Logical argument indicating whether to drop unused depths
from the legend. Default is \code{legend_drop = FALSE}}

\item{legend_position}{Position of the legend ("none", "left", "right",
"bottom", "top"). Default is \code{legend_position = "right"}.}

\item{ylims}{Numeric vector of limits for the y-axis.}

\item{date_axis_name}{Name for the x-axis.}

\item{date_breaks_major}{Character string specifying intervals for major
breaks on the x-axis, e.g., \code{date_breaks_major = "2 month"}. Default
is \code{date_breaks_major = NULL}, which chooses breaks based on the
length of the data series.}

\item{date_breaks_minor}{Character string specifying intervals for major
breaks on the x-axis, e.g., \code{date_breaks_minor = "1 month"}. Default
is \code{date_breaks_minor = NULL}, which chooses breaks based on the
length of the data series.}

\item{date_labels_format}{Character string specifying the format of the date
labels on the x-axis, e.g., \code{date_labels_format = "\%y-\%b"}. Default
is \code{date_labels_format = NULL}, which chooses labels based on the
length of the data series.}

\item{plotly_friendly}{Logical argument. If TRUE, y-axis label is set to a
plotly-friendly title ("Temperature (deg C)"). If FALSE,
\code{expression()} is used to insert the degree symbol.}

\item{alpha}{Transparency for the heat stress and superchill shaded boxes.}
}
\value{
Returns a ggplot object.
}
\description{
Plot filtered data with removed data greyed out
}
