% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature_at_depth.R
\name{plot_temperature_at_depth}
\alias{plot_temperature_at_depth}
\title{Plot temperature coloured by depth}
\usage{
plot_temperature_at_depth(
  dat,
  trend_threshold = 4,
  superchill_threshold = -0.7,
  heat_threshold = 18,
  facet_var = NULL,
  ncol = 1,
  nrow = NULL
)
}
\arguments{
\item{dat}{Dataframe with at least three columns: \code{TIMESTAMP} (POSIXct
or character in the order year, month, day, hours, minutes, seconds),
\code{DEPTH}, and \code{VALUE}. If column \code{VARIABLE} is included, it
must have one unique entry. Other columns will be ignored.}

\item{trend_threshold}{Default is \code{trend_threshold = 4}. The last
observation above \code{trend_threshold} that does not return below
\code{trend_threshold} triggers the beginning of the growing season for
each group in \code{...}.}

\item{superchill_threshold}{Default is \code{superchill_threshold = -0.7}.
The first observation below \code{superchill_threshold} triggers the end of
the growing season for each group in \code{...}.}

\item{heat_threshold}{The threshold to trigger heat stress interval
(inclusive. Default is \code{heat_threshold = 18}).}

\item{facet_var}{Variable(s) defining faceting groups. Variables must be
column(s) in \code{dat}). For a single facet variable: \code{facet_var =
  "SEASON"}. For more than one facet variables: \code{facet_var = "SEASON +
  DEPTH"}. Default is \code{facet_var = NULL}.}

\item{ncol}{Number of columns for faceted figure. Default is \code{ncol = 1}.}

\item{nrow}{Number of rows for faceted figure. Default is \code{nrow = NULL}.}
}
\value{
ggplot object
}
\description{
Plot temperature coloured by depth
}
