% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgc_functions.R
\name{TGC_calculate_final_weight}
\alias{TGC_calculate_final_weight}
\title{Calculate final body weight using TGC model}
\usage{
TGC_calculate_final_weight(dd_table, initial_weight, tgc)
}
\arguments{
\item{dd_table}{Dataframe with degree-day data, as exported from
\code{count_degree_days}. Must include columns \code{n_degree_days}.}

\item{initial_weight}{A vector of initial weight(s) of the fish in grams.}

\item{tgc}{A vector of thermal growth coefficient(s).}
}
\value{
Returns \code{dd_table} additional columns \code{INITIAL_WEIGHT},
\code{TGC}, and \code{TGC_FINAL_WEIGHT}.
}
\description{
Calculate final body weight using TGC model
}
\details{
Final weight is calculated from the thermal growth coefficient (TGC)
model:

\deqn{final_weight = (initial_weight^(1/3) + (TGC/100)*degree_days)^3 }

where \eqn{final_weight} and \eqn{initial_weight} are in grams.

Final weight will be calculated for all combinations of
\code{initial_weight} and \code{tgc} supplied in the arguments and each row
of \code{dd_table}.
}
